# perf/integration tests: Preparing for precalculated wallet addresses lib feature