A new feature on the wallet-lib will allow wallets to have their addresses pre-calculated to save up on processing time at the integration tests.

This PR contains the code adaptations to use this feature, as designed by:
- #173 

Once the wallet lib is updated with the address pre-calculation feature, the following average time improvements are expected:
- Common wallets start time: from 9 seconds to 2.3 seconds
- Multisig wallets start time: from 58 seconds to 4.6 seconds
More details on this available on the 173 design.

With this change alone, runs will be executed in around 500 seconds, a gain of near 50% from the original ~1000 seconds.
The lib version bump will be made in a separate PR.

### Summary of changes
- Adds the _pre calculated wallet addresses_ feature
  - Script for generating these wallets
  - Helper class for the main wallet logic
  - Refactored `WalletHelper` to use it
  - Mocked tests to cover the new code
- Created benchmark classes to measure time in a more readable way:
  - Wallet initialization class
  - Transaction execution class
  - The `test-utils-integration.js` methods were refactored to use them
  - The `WalletHelper.startMultipleWalletsForTest` method was simplified
- Refactored test constants into their own file
  - Some mocked tests started needing access to this constant data
- Refactored common test functions into a `core.util.js` file
  - Avoid circular dependency in some cases
- Developer scripts moved from `src/utils` to `scripts/`
  - Both for structure clarity and for keeping test coverages correct

### Acceptance Criteria
- All tests should pass with the current version of the wallet lib
- All integration tests should use the new wallet pre-calculation feature

### Security Checklist
- [X] Make sure you do not include new dependencies in the project unless strictly necessary and do not include dev-dependencies as production ones. More dependencies increase the possibility of one of them being hijacked and affecting us.