@tuliomir 2022-05-02
About the utils, I agree they should be moved to a scripts folder. This was done on [ed0d3e1](https://github.com/HathorNetwork/hathor-wallet-headless/pull/189/commits/ed0d3e1f443dcd4c6ff54e0658c8bd74f80c5bee).

About the precalculation helper, that is still on the `src` folder, I'm having mixed ideas about what to do with it. I have a feeling it _might_ be used on the application in the future, as some kind of helper lib or simulator actionable via http request.
- If this proves to be true, it would need tests. I erred on the side of caution and made them already.
- If it keeps being only a script helper, it could also be moved to the `/scripts` folder, along with its tests.